feat: expo dev client
