+++ ausweis-example/package.json
@@ -3,15 +3,17 @@
   "main": "expo-router/entry",
   "version": "1.0.0",
   "scripts": {
-    "start": "expo start",
+    "start": "expo start -c --dev-client",
     "android": "expo run:android",
-    "ios": "expo run:ios"
+    "ios": "expo run:ios",
+    "prebuild": "expo prebuild --no-install"
   },
   "dependencies": {
     "@animo-id/expo-ausweis-sdk": "../",
     "@react-navigation/native": "^6.0.2",
     "expo": "~51.0.17",
     "expo-build-properties": "~0.12.3",
+    "expo-dev-client": "~4.0.20",
     "expo-router": "~3.5.17",
     "react": "18.2.0",
     "react-native": "0.74.2",
